Connect command divided into three connectors
